requirements.txt
fastapi
uvicorn[standard]
gunicorn
supabase
openai
anthropic
google-generativeai
python-telegram-bot
requests
pydantic
sqlalchemy[asyncio]
asyncpg
slowapi
python-dotenv
sqlalchemy
psycopg2-binary
python-dotenv
httpx
python-jose
passlib[bcrypt]